Rearrange helper functions
