+++ e2e_test/__init__.py
@@ -1,0 +1,3 @@
+"""
+E2E test package for Copy as Markdown extension.
+"""
